CoreData helper
CoreData helper is a class that handles everything CoreData / data persistence to the project. Two initializers based on what you're attempting to do. 1 ) is for general use, such as history page. 2 ) is for saving an articles states.

Functions - attemptUpdate - checks if a save exists, if it does it updates, otherwise it saves a new save.
saveArticle - saves an article to core data.
deleteAll is a temp function for wiping all data.
loadArticles loads all articles saved.